Scripts/Daniel/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2        
chromedriver-autoinstaller==0.6.4
cryptography==42.0.7
cssselect==1.2.0
Deprecated==1.2.14
et-xmlfile==1.1.0
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
jmespath==1.0.1
lxml==5.2.2
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.21.0
sidrapy==0.1.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
